requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anndata=0.9.1=pyhd8ed1ab_0
anyio=3.5.0=py39h06a4308_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
arpack=3.7.0=hdefa2d7_2
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=py39h06a4308_0
babel=2.11.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
comm=0.1.2=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glpk=5.0=h445213a_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
h5py=3.7.0=py39h737f45e_0
hdf5=1.10.6=nompi_h6a2412b_1114
icu=72.1=hcb278e6_0
idna=3.4=py39h06a4308_0
igraph=0.10.4=hb9ddf80_2
importlib-metadata=6.0.0=py39h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py39hb070fc8_0
ipython=8.12.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.2=py39h06a4308_0
joblib=1.1.1=py39h06a4308_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py39h06a4308_0
jupyter_client=8.1.0=py39h06a4308_0
jupyter_core=5.3.0=py39h06a4308_0
jupyter_server=1.23.4=py39h06a4308_0
jupyterlab=3.5.3=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.22.0=py39h06a4308_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
leidenalg=0.9.1=py39h227be39_0
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=12_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcurl=7.87.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libiconv=1.17=h166bdaf_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h9e868ea_6
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxml2=2.10.4=hfdac1af_0
libxslt=1.1.37=h873f0b0_0
libzlib=1.2.13=h166bdaf_4
lightning-utilities=0.8.0=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.39.1=py39he621ea3_0
lxml=4.9.2=py39h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.1=py39h417a72b_1
matplotlib-inline=0.1.6=py39h06a4308_0
metis=5.1.0=hf484d3e_4
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
multidict=6.0.4=pypi_0
munkres=1.1.4=py_0
natsort=7.1.1=pyhd3eb1b0_0
nbclassic=0.5.5=py39h06a4308_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.5.4=py39h06a4308_0
nbformat=5.7.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
notebook=6.5.4=py39h06a4308_0
notebook-shim=0.2.2=py39h06a4308_0
numba=0.56.4=py39h71a7301_1
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h0b41bf4_0
packaging=23.0=py39h06a4308_0
pandas=2.0.1=py39h40cae4c_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
patsy=0.5.3=py39h06a4308_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
platformdirs=2.5.2=py39h06a4308_0
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=4.22.3=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyg-lib=0.2.0+pt20cu118=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pynndescent=0.5.4=pyhd3eb1b0_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python-igraph=0.10.4=py39hfac54ea_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=2.0.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_3
pytorch-lightning=2.0.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=py39h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.29.0=py39h06a4308_0
scanpy=1.9.3=pyhd8ed1ab_0
scikit-learn=1.2.2=py39h6a678d5_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=py39h06a4308_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.21.1=pypi_0
session-info=1.0.0=pyhd8ed1ab_0
setproctitle=1.3.2=pypi_0
setuptools=66.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py39h06a4308_1
soupsieve=2.4=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.5=py39h7deecbd_1
stdlib-list=0.8.0=py39h06a4308_0
suitesparse=5.10.1=h446ee2e_2
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
terminado=0.17.1=py39h06a4308_0
texttable=1.6.4=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py39h06a4308_0
torch-cluster=1.6.1+pt20cu118=pypi_0
torch-geometric=2.3.1=pypi_0
torch-scatter=2.1.1+pt20cu118=pypi_0
torch-sparse=0.6.17+pt20cu118=pypi_0
torch-spline-conv=1.2.2+pt20cu118=pypi_0
torchmetrics=0.11.4=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu118
tornado=6.2=py39h5eee18b_0
tqdm=4.65.0=py39hb070fc8_0
traitlets=5.7.1=py39h06a4308_0
typing-extensions=4.5.0=py39h06a4308_0
typing_extensions=4.5.0=py39h06a4308_0
tzdata=2023c=h04d1e81_0
umap-learn=0.5.3=py39hf3d152e_0
urllib3=1.26.15=py39h06a4308_0
wandb=0.15.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
wheel=0.38.4=py39h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.10=h5eee18b_1
yarl=1.9.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.5=hc292b87_0
